requirements.txt
pydantic-ai
groq
python-telegram-bot[job-queue]
tweepy
python-dotenv
loguru
pytz
fastapi
uvicorn
jinja2
ddgs

nest-asyncio
logfire

langchain-community 
wikipedia
pydantic-ai-slim[tavily]